Use a single page. Menu items point to sections.
Each section contains a heading element. If this element contains a link,
that link displays a paragraph symbol when hovered (like in some other
websites).

The sections linked from the top navigation menu contain a link to go
back to the top.

Every section can be linked. So it should be easier to share with others.

Moved the part about previous conferences to footer as that seems less
relevant (not about the currency CWLConf 2024).