Drop all headers other than home to 'h3'
Call for submissions to h4 (consider having an 'announcements' section that's h3 and call for submissions goes under that?

Updated section breaks to 4 rem, for small content sections (like project and directions) this looks a little off for now, but those sections will grow.

I like having the previous conference links in the footnotes, but was still in list format on a single line, instead use a '/' to separate each link